requirements.txt
python==3.8.2
python-dotenv==1.0.1
git-python==1.0.3
langchain-community==0.2.0
langchain-text-splitters==0.2.0
langchain==0.2.0
langchain-core==0.2.0
langchain-openai==0.1.7
langchain-chroma==0.1.1